backend/requirements.txt
Flask
Flask-Login
Flask-Migrate
Flask-SQLAlchemy
Flask-CORS
python-dotenv
Werkzeug
patchright 